Adjustments to the input signal
